requirements.txt
Flask==2.2.3
Flask_RESTful==0.3.9
flask_sqlalchemy==3.0.3
flask_swagger_ui==4.11.1
pandas==1.3.5
psycopg2_binary==2.9.6
SQLAlchemy==1.4.46
